requirements.txt
groq
google-generativeai
python-dotenv
requests
openai
cohere